Finally a version working after multiple bug fixes.

+++ lib/include/ppu/oam_data.h
@@ -4,15 +4,15 @@
 namespace gameboy::ppu {
 struct OAMEntry
 {
-    int8_t y;
-    int8_t x;
-    int8_t tile;
+    uint8_t y;
+    uint8_t x;
+    uint8_t tile;
 
-    unsigned f_cgb_palette : 3;
-    unsigned f_cgb_vram_bank : 1;
-    unsigned f_dmg_palette : 1;
-    unsigned f_x_flip : 1;
-    unsigned f_y_flip : 1;
-    unsigned f_background_priority : 1;
+    uint8_t f_cgb_palette : 3;
+    uint8_t f_cgb_vram_bank : 1;
+    uint8_t f_dmg_palette : 1;
+    uint8_t f_x_flip : 1;
+    uint8_t f_y_flip : 1;
+    uint8_t f_background_priority : 1;
 };
 }  // namespace gameboy::ppu
